Adding filter to properties
